# book/docs/swagger.yaml
info:
  contact: {}
paths:
  /admin_list:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: size
        in: query
        name: size
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: phone
        in: query
        name: phone
        type: string
      - description: email
        in: query
        name: email
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 管理员列表
      tags:
      - 管理员方法
  /book_create:
    post:
      parameters:
      - description: name
        in: formData
        name: name
        required: true
        type: string
      - description: description
        in: formData
        name: description
        required: true
        type: string
      - description: publish_date
        in: formData
        name: publish_date
        required: true
        type: string
      - description: author
        in: formData
        name: author
        required: true
        type: string
      - description: publisher
        in: formData
        name: publisher
        required: true
        type: string
      - description: bookNo
        in: formData
        name: bookNo
        required: true
        type: string
      - description: score
        in: formData
        name: score
        required: true
        type: string
      - description: nums
        in: formData
        name: nums
        required: true
        type: string
      - description: category_id
        in: formData
        name: category_id
        required: true
        type: string
      - description: cover
        in: formData
        name: cover
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 图书创建
      tags:
      - 图书方法
  /book_delete:
    delete:
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 图书删除
      tags:
      - 图书方法
  /book_detail:
    get:
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 图书详细
      tags:
      - 图书方法
  /book_list:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: size
        in: query
        name: size
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: id
        in: query
        name: id
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 图书列表
      tags:
      - 图书方法
  /book_update:
    put:
      parameters:
      - description: id
        in: formData
        name: id
        required: true
        type: integer
      - description: name
        in: formData
        name: name
        type: string
      - description: description
        in: formData
        name: description
        type: string
      - description: publish_date
        in: formData
        name: publish_date
        type: string
      - description: author
        in: formData
        name: author
        type: string
      - description: publisher
        in: formData
        name: publisher
        type: string
      - description: bookNo
        in: formData
        name: bookNo
        type: string
      - description: score
        in: formData
        name: score
        type: string
      - description: nums
        in: formData
        name: nums
        type: string
      - description: category_id
        in: formData
        name: category_id
        type: string
      - description: cover
        in: formData
        name: cover
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 图书修改
      tags:
      - 图书方法
  /bookimgup:
    post:
      parameters:
      - description: token
        in: query
        name: token
        required: true
        type: string
      - description: file
        in: formData
        name: file
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 头像上传
      tags:
      - 图书方法
  /borrow_create:
    post:
      parameters:
      - description: book_id
        in: formData
        name: book_id
        required: true
        type: string
      - description: user_id
        in: formData
        name: user_id
        required: true
        type: string
      - description: days
        in: formData
        name: days
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200","msg":""}'
          schema:
            type: string
      summary: 新建借书
      tags:
      - 借书方法
  /borrow_delete:
    delete:
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 借书删除
      tags:
      - 借书方法
  /borrow_list:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: size
        in: query
        name: size
        type: string
      - description: user_name
        in: query
        name: user_name
        type: string
      - description: book_name
        in: query
        name: book_name
        type: string
      - description: book_id
        in: query
        name: book_id
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 借书列表
      tags:
      - 借书方法
  /category_create:
    post:
      parameters:
      - description: name
        in: formData
        name: name
        required: true
        type: string
      - description: remark
        in: formData
        name: remark
        type: string
      responses:
        "200":
          description: '{"code":"200","msg":""}'
          schema:
            type: string
      summary: 添加分类
      tags:
      - 分类方法
  /category_delete:
    delete:
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 分类删除
      tags:
      - 分类方法
  /category_detail:
    get:
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 分类详细
      tags:
      - 分类方法
  /category_list:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: size
        in: query
        name: size
        type: string
      - description: name
        in: query
        name: name
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 分类列表
      tags:
      - 分类方法
  /category_update:
    put:
      parameters:
      - description: id
        in: formData
        name: id
        required: true
        type: integer
      - description: name
        in: formData
        name: name
        type: string
      - description: remark
        in: formData
        name: remark
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 分类修改
      tags:
      - 分类方法
  /code:
    post:
      parameters:
      - description: email
        in: formData
        name: email
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 发送邮箱验证码
      tags:
      - 管理员方法
  /echart_data:
    get:
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: echarts数据
      tags:
      - echarts数据
  /login:
    post:
      parameters:
      - description: name
        in: formData
        name: name
        type: string
      - description: password
        in: formData
        name: password
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 管理员登录
      tags:
      - 管理员方法
  /register:
    post:
      parameters:
      - description: name
        in: formData
        name: name
        required: true
        type: string
      - description: password
        in: formData
        name: password
        required: true
        type: string
      - description: password2
        in: formData
        name: password2
        required: true
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      - description: email
        in: formData
        name: email
        required: true
        type: string
      - description: code
        in: formData
        name: code
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 管理员注册
      tags:
      - 管理员方法
  /retur_create:
    post:
      parameters:
      - description: book_id
        in: formData
        name: book_id
        required: true
        type: string
      - description: user_id
        in: formData
        name: user_id
        required: true
        type: string
      - description: days
        in: formData
        name: days
        required: true
        type: string
      - description: id
        in: formData
        name: id
        required: true
        type: string
      - description: return_date
        in: formData
        name: return_date
        required: true
        type: string
      - description: created_at
        in: formData
        name: created_at
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 还书创建
      tags:
      - 还书方法
  /retur_delete:
    delete:
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 还书删除
      tags:
      - 还书方法
  /retur_list:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: size
        in: query
        name: size
        type: string
      - description: user_name
        in: query
        name: user_name
        type: string
      - description: book_name
        in: query
        name: book_name
        type: string
      - description: book_id
        in: query
        name: book_id
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 还书列表
      tags:
      - 还书方法
  /user_account:
    post:
      parameters:
      - description: id
        in: formData
        name: id
        required: true
        type: string
      - description: score
        in: formData
        name: score
        required: true
        type: integer
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 账号充值
      tags:
      - 用户方法
  /user_create:
    post:
      parameters:
      - description: name
        in: formData
        name: name
        required: true
        type: string
      - description: age
        in: formData
        name: age
        required: true
        type: string
      - description: sex
        in: formData
        name: sex
        required: true
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      - description: address
        in: formData
        name: address
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200","msg":""}'
          schema:
            type: string
      summary: 用户注册
      tags:
      - 用户方法
  /user_delete:
    delete:
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 用户删除
      tags:
      - 用户方法
  /user_detail:
    get:
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 用户详细
      tags:
      - 用户方法
  /user_list:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: size
        in: query
        name: size
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: phone
        in: query
        name: phone
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 用户列表
      tags:
      - 用户方法
  /user_update:
    put:
      parameters:
      - description: id
        in: formData
        name: id
        required: true
        type: integer
      - description: name
        in: formData
        name: name
        type: string
      - description: age
        in: formData
        name: age
        type: string
      - description: address
        in: formData
        name: address
        type: string
      - description: sex
        in: formData
        name: sex
        type: string
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 用户修改
      tags:
      - 用户方法
swagger: "2.0"